Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -9,10 +9,10 @@
  */
 int	main(void)
 {
-	printf("Size of a char: %ld byte(s)\n", sizeof(char));
-	printf("Size of an int: %ld byte(s)\n", sizeof(int));
-	printf("Size of long int: %ld byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %ld byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %ld byte(s)\n", sizeof(float));
+	printf("Size of a char: %d byte(s)\n", (int)sizeof(char));
+	printf("Size of an int: %d byte(s)\n", (int) sizeof(int));
+	printf("Size of long int: %d byte(s)\n", (int) sizeof(long int));
+	printf("Size of a long long int: %d byte(s)\n", (int) sizeof(long long int));
+	printf("Size of a float: %d byte(s)\n", (int)sizeof(float));
 	return (0);
 }
